+++ crates/full-node/sov-db/src/storage_manager/nomt_based/groups.rs
@@ -10,7 +10,6 @@ use anyhow::Context;
 use rockbound::cache::delta_reader::DeltaReader;
 use rockbound::versioned_db::VersionedDeltaReader;
 use rockbound::SchemaBatch;
-#[cfg(feature = "migration-script")]
 use sov_rollup_interface::common::SlotNumber;
 use sov_rollup_interface::reexports::digest;
 
@@ -75,6 +74,14 @@ where
     }
 
     pub(crate) fn commit(&mut self, group: CommitGroup) -> anyhow::Result<()> {
+        self.commit_helper(group, None)
+    }
+
+    fn commit_helper(
+        &mut self,
+        group: CommitGroup,
+        expected_latest: Option<SlotNumber>,
+    ) -> anyhow::Result<()> {
         tracing::trace!("Commiting a group...");
         // The last commit had to be successful.
         let CommitGroup {
@@ -106,7 +113,22 @@ where
 
         // Flat State
         tracing::trace!("Committing Flat DB..");
-        let flat_metrics = self.flat_state.commit(historical_state)?;
+        #[cfg(feature = "migration-script")]
+        let flat_metrics = if let Some(expected_latest) = expected_latest {
+            self.flat_state
+                .commit_at_latest_checked(historical_state, expected_latest)?
+        } else {
+            self.flat_state.commit(historical_state)?
+        };
+
+        #[cfg(not(feature = "migration-script"))]
+        let flat_metrics = {
+            assert!(
+                expected_latest.is_none(),
+                "expected_latest must be none when not in migration script"
+            );
+            self.flat_state.commit(historical_state)?
+        };
 
         // Metrics
         let merklized_commit_from_caller = merklized_commit.total;
@@ -131,46 +153,7 @@ where
         group: CommitGroup,
         expected_latest: SlotNumber,
     ) -> anyhow::Result<()> {
-        tracing::trace!(%expected_latest, "Committing a group at latest with expected-head check...");
-        let CommitGroup {
-            nomt: state,
-            rockbound:
-                SnapshotGroup {
-                    historical_state,
-                    accessory,
-                    ledger,
-                },
-        } = group;
-
-        tracing::trace!("Commiting NOMT DBs...");
-        let merklized_commit = self.merklized_state.commit(state)?;
-
-        tracing::trace!("Committing Ledger DB...");
-        let ledger_commit = self.commit_ledger(&ledger)?;
-
-        tracing::trace!("Commiting Accessory DB...");
-        let accessory_commit =
-            self.commit_accessory(&accessory, &historical_state.root_hash_batch)?;
-
-        tracing::trace!("Committing Flat DB at latest version...");
-        let flat_metrics = self
-            .flat_state
-            .commit_at_latest_checked(historical_state, expected_latest)?;
-
-        let merklized_commit_from_caller = merklized_commit.total;
-        let commit_detailed_metrics = CommitDetailedMetric {
-            merklized_commit,
-            merklized_commit_from_caller,
-            flat: flat_metrics,
-            accessory_commit,
-            ledger_commit,
-        };
-        sov_metrics::track_metrics(|tracker| {
-            tracker.submit(commit_detailed_metrics);
-        });
-        self.merklized_state.send_metrics();
-
-        Ok(())
+        self.commit_helper(group, Some(expected_latest))
     }
 
     #[cfg(feature = "migration-script")]